+++ test/javascript/components/mentoring/mentor-discussion/Guidance.test.js
@@ -2,7 +2,7 @@ import React from 'react'
 import userEvent from '@testing-library/user-event'
 import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
-import { Guidance } from '../../../../app/javascript/components/mentoring/Guidance'
+import { Guidance } from '../../../../../app/javascript/components/mentoring/mentor-discussion/Guidance'
 
 test('how you solved the exercise is open by default', async () => {
   render(<Guidance />)
